requirements.txt
mcp>=1.6.0
google-analytics-data>=0.20.0
google-analytics-admin>=0.27.0
google-auth>=2.40.0
httpx>=0.28.1
python-dotenv>=1.0.0
cryptography>=44.0.0
uvicorn[standard]>=0.32.0
